+++ visualization/script/travis-build.sh
@@ -3,6 +3,7 @@
 cd visualization
 npm install
 npm run build
-npm run test
+npm run test --ci
+npm run e2e --ci
 npm run doc
 npm run package