+++ analysis/model/src/main/kotlin/de/maibornwolff/codecharta/model/DependencyType.kt
@@ -27,38 +27,8 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-package de.maibornwolff.codecharta.serialization
+package de.maibornwolff.codecharta.model
 
-import org.hamcrest.MatcherAssert.assertThat
-import org.hamcrest.Matchers.*
-import org.jetbrains.spek.api.Spek
-import org.jetbrains.spek.api.dsl.it
-import java.io.StringReader
-
-class ProjectDeserializerTest : Spek({
-    val EXAMPLE_CC_JSON = "example.cc.json"
-
-    it("shouldDeserializeProjectJson") {
-        val expectedJsonReader = this.javaClass.classLoader.getResourceAsStream(EXAMPLE_CC_JSON).reader()
-
-        val project = ProjectDeserializer.deserializeProject(expectedJsonReader)
-
-        assertThat(project.projectName, `is`("201701poolobject"))
-        assertThat(project.size, `is`(6))
-    }
-
-    it("deserializeProject_should_map_nonexisting_values_to_defaults") {
-        // given
-        val jsonString = "{projectName='some Project', apiVersion='1.0', nodes=[{name:'root',type:'Folder'}]}"
-
-        // when
-        val project = ProjectDeserializer.deserializeProject(StringReader(jsonString))
-
-        // then
-        val node = project.rootNode
-
-        assertThat(node.link, nullValue())
-        assertThat(node.attributes, not(nullValue()))
-        assertThat(node.children, not(nullValue()))
-    }
-})
+enum class DependencyType {
+    static, temporal_coupling
+}