+++ analysis/model/src/main/kotlin/de/maibornwolff/codecharta/model/Project.kt
@@ -33,7 +33,8 @@ package de.maibornwolff.codecharta.model
 class Project(
         val projectName: String,
         private val nodes: List<Node> = listOf(Node("root", NodeType.Folder)),
-        val apiVersion: String = API_VERSION
+        val apiVersion: String = API_VERSION,
+        val dependencies: MutableMap<DependencyType, MutableList<Dependency>> = mutableMapOf()
 ) {
     init {
         if (nodes.size != 1) throw IllegalStateException("no root node present in project")
@@ -45,16 +46,33 @@ class Project(
     val size: Int
         get() = rootNode.size
 
+    fun sizeOfDependencies(dependencyType: DependencyType): Int {
+        if (dependencies.get(dependencyType) !== null) {
+            return dependencies.get(dependencyType)!!.size
+        } else {
+            return 0
+        }
+
+    }
+
     override fun toString(): String {
         return "Project{" +
                 "projectName='" + projectName + '\''.toString() +
                 ", apiVersion='" + apiVersion + '\''.toString() +
                 ", nodes=" + nodes +
+                ", dependencies=" + dependencies +
                 '}'.toString()
     }
 
     companion object {
-        const val API_VERSION = "1.0"
+        private const val API_VERSION_MAJOR = "1"
+        private const val API_VERSION_MINOR = "1"
+        const val API_VERSION = "$API_VERSION_MAJOR.$API_VERSION_MINOR"
+
+        fun isAPIVersionCompatible(apiVersion: String): Boolean {
+            val apiVersion_major = apiVersion.split('.')[0]
+            return apiVersion_major == API_VERSION_MAJOR
+        }
     }
 
 }