+++ analysis/tools/ccsh/src/main/kotlin/de/maibornwolff/codecharta/tools/ccsh/Ccsh.kt
@@ -31,6 +31,7 @@ package de.maibornwolff.codecharta.tools.ccsh
 
 import de.maibornwolff.codecharta.exporter.csv.CSVExporter
 import de.maibornwolff.codecharta.filter.mergefilter.MergeFilter
+import de.maibornwolff.codecharta.importer.codemaat.CodeMaatCouplingImporter
 import de.maibornwolff.codecharta.importer.csv.CSVImporter
 import de.maibornwolff.codecharta.importer.csv.SourceMonitorImporter
 import de.maibornwolff.codecharta.importer.scmlogparser.SCMLogParser
@@ -54,7 +55,8 @@ import java.util.concurrent.Callable
             Installer::class,
             CSVExporter::class,
             CrococosmoImporter::class,
-            UnderstandImporter::class
+            UnderstandImporter::class,
+            CodeMaatCouplingImporter::class
         ],
         versionProvider = Ccsh.ManifestVersionProvider::class,
         footer = ["Copyright(c) 2018, MaibornWolff GmbH"]