+++ visualization/app/codeCharta/ui/fileChooser/fileChooserController.ts
@@ -84,10 +84,10 @@ class FileChooserController {
      */
     setNewData(name, parsedData, revision){
         let ctx = this;
-        this.dataLoadingService.loadMapFromFileContent(name, parsedData, revision).then(
+        return this.dataLoadingService.loadMapFromFileContent(name, parsedData, revision).then(
             () => {
 
-                ctx.scenarioService.applyScenario(this.scenarioService.getDefaultScenario());
+                ctx.scenarioService.applyScenarioOnce(this.scenarioService.getDefaultScenario());
                 ctx.dataService.setComparisonMap(revision);
                 ctx.dataService.setReferenceMap(revision);